+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [3.3.0] - 19-08-2026
 
+### Fixed
+- Fallback sendOnBehalfUser failed. Change from `Get-Mailbox` to `Get-EXOMailbox` and added `-RecipientTypeDetails UserMailbox`.
+
 ### Removed
 - Removed Client Secret support from the repository. Make sure to also update `configuration.json` when updating the connector.